core: deprecate attr_constr_coercion
